export IsbnData to a typing file to make it available to client
