+++ types/common.ts
@@ -35,3 +35,10 @@ export type HighResolutionTime = [ number, number ]
 export type ISODate = string
 
 export type StringifiedHashedSecretData = `{${string}}`
+
+export type IsbnData = ReturnType<typeof isbnParser> & {
+  groupPrefix?: string
+  publisherPrefix?: string
+  groupLang?: string
+  groupLangUri?: string
+}
